+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 - `missing-input-background`: (new) Missing input labels get a red background coloring
 
+### Fixed
+
+- Fix incorrect bolding of command arguments in context controls
+
 ## 25.2.27
 
 ### Fixed